tap-news-backendserver/requirements.txt
jsonrpclib-pelix
pika
pymongo